pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "arc-mvp-skeleton"
version = "0.1.0"
description = "MVP skeleton for the arc archival API and CLIs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "OpenAI" }]
dependencies = [
  "fastapi>=0.111,<1.0",
  "pydantic>=2.7,<3.0",
  "httpx>=0.27,<1.0",
  "typer>=0.12,<1.0",
  "uvicorn>=0.30,<1.0",
  "PyYAML>=6.0,<7.0",
]


[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5,<1.0",
  "mypy>=1.10,<2.0",
]

planner = [
  "numpy>=2.0,<3.0",
  "scipy>=1.13,<2.0",
]

db = [
  "sqlalchemy>=2.0,<3.0",
]

[project.scripts]
arc-api = "arc_api.app:main"
arc = "arc_cli.main:main"
arc-disc = "arc_disc.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/arc_api", "src/arc_cli", "src/arc_core", "src/arc_disc"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["arc_api", "arc_cli", "arc_core", "arc_disc"]
